clean up config hook
